根据级别查询

+++ back_manage_b/src/main/java/com/zhs/model/vo/ResumeLevelAreaVO.java
@@ -9,7 +9,7 @@ import org.aspectj.lang.annotation.DeclareAnnotation;
  */
 @Data
 public class ResumeLevelAreaVO {
-    private Integer areaId;
+    private Long areaId;
     private String areaName;
     private Integer resumeCount;
 }
